feat: remove mydumper from dm in build-common

+++ jenkins/pipelines/cd/atom-jobs/build-common.groovy
@@ -609,7 +609,7 @@ if [[ -d "dm/dm/dm-ansible" ]]; then
     mv dm/monitoring ${TARGET}/
 fi;
 
-if [[ ${ARCH} == "amd64" ]]; then
+if [[ ${ARCH} == "amd64" && ${RELEASE_TAG} < "v7.5.0" ]]; then
     curl http://download.pingcap.org/mydumper-latest-linux-amd64.tar.gz | tar xz
     mv mydumper-latest-linux-amd64/bin/mydumper bin/ && rm -rf mydumper-latest-linux-amd64
 fi;
